third_party/WebKit/LayoutTests/http/tests/security/create-document-change-domain.html

<!DOCTYPE html>
<html>
<head>
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
    <script>
        if (document.location.hostname == "127.0.0.1") {
            document.location.hostname = "subdomain.example.test";
        } else {
            test(function () {
                var doc = document.implementation.createDocument(null, 'one', null);

                assert_equals(document.domain, 'subdomain.example.test');
                assert_equals(doc.domain, 'subdomain.example.test');

                document.domain = 'example.test';
                assert_equals(document.domain, 'example.test');
                assert_equals(doc.domain, 'example.test');

                document.domain = 'subdomain.example.test';
            }, "Documents created using createDocument share the underlying domain.");
        }
    </script>
</head>
<body>
</body>
</html>
